@@ -1,6 +1,6 @@
-import numpy as np
 from collections import defaultdict
 
+import numpy as np
 import tlc
 
 
